% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhenotype.R
\name{getPhenotype}
\alias{getPhenotype}
\title{Phenotype labelling}
\usage{
getPhenotype(clusterID, cellExpr)
}
\arguments{
\item{clusterID}{character vector with identifier for each cell to a cluster}

\item{cellExpr}{data.frame with numerical cell expression. Rows represents
the cells that correspond with the clusterID. Columns represent the markers
that will be used for phenotyping.}
}
\value{
a list, check description
}
\description{
returns a list with three elements `$phenotype_table`, `$marker_bins` and
`$marker_quantiles`.

* phenotype_table, a `matrix`. Rows represent the clusterID and columns
  represent the markers. Each entry is either 1 (positive), -1 (negative) or 0 (either)

* marker_bins, a `list` with two elements, the $background and $clusters specific
  bins

* marker_quantiles, a `list` with n (number of markers) elements. Each element
  is a matrix with 41 rows (quantiles of 2.5%) and p (number of clusters). Thus each
  entry is a specific quantile of a specific cluster for the given marker.
}
\details{
I'll explain the algorithm later
}
\keyword{labels}
\keyword{markers,}
\keyword{phenotype,}
