% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizeMarkers.R
\name{harmonizeMarkers}
\alias{harmonizeMarkers}
\title{Harmonize cellular marker labels}
\usage{
harmonizeMarkers(markers)
}
\arguments{
\item{markers}{character}
}
\value{
character
}
\description{
Harmonization of marker labels. Each element of a given character
vector will be evaluated and assigned a common label.
}
\keyword{labels}
\keyword{markers,}
\keyword{panel,}
