% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{label_list}
\alias{label_list}
\title{label_list}
\format{
## `label_list`
List with gating scheme on how to label the phenotypes generated by getPhenotype
}
\usage{
label_list
}
\description{
label_list
}
\keyword{datasets}
