% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelPhenotype.R
\name{labelPhenotype}
\alias{labelPhenotype}
\title{Phenotype labeller}
\usage{
labelPhenotype(phenotype_table)
}
\arguments{
\item{labels}{character}
}
\value{
list
}
\description{
Function that labels the phenotype table generated by \code{\link[CyStore]{getPhenotype}}
}
\keyword{labels,}
\keyword{markers,}
\keyword{phenotype,}
\keyword{scheme}
