% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateToPanel.R
\name{templateToPanel}
\alias{templateToPanel}
\title{Draft panel from CyTOF template}
\usage{
templateToPanel(template)
}
\arguments{
\item{template}{character, location of file}
}
\value{
data.frame
}
\description{
Function that takes a template file used for the CyTOF. It will
convert this template to a data frame containing all relevant information.
It will also add a column with harmonized markers see
\link[CyStore]{harmonizeMarkers.R} for more information
}
\keyword{cytof}
\keyword{labels,}
\keyword{markers,}
\keyword{panel,}
\keyword{template,}
